Adding auth to server. Deleting bookman-old folder.
