+++ app.js
@@ -1,32 +1,84 @@
-/**
- * Module dependencies
- */
+// Express 
+var express = require('express');
+var errorHandler = require('express-error-handler');
+var session = require('express-session')
 
-var express = require('express'),
-  bodyParser = require('body-parser'),
-  methodOverride = require('method-override'),
-  errorHandler = require('express-error-handler'),
-  morgan = require('morgan'),
-  routes = require('./routes'),
-  api = require('./routes/api'),
-  http = require('http'),
-  path = require('path');
+// Use dependencies
+var morgan = require('morgan');
+var bodyParser = require('body-parser');
+var methodOverride = require('method-override');
+
+var routes = require('./routes');
+var api = require('./routes/api');
+var http = require('http');
+var path = require('path');
+
+// Env
+var dotenv = require('dotenv');
+dotenv.load();
+
+// Auth 
+var passport = require('passport');
+var GoodreadsStrategy = require('passport-goodreads').Strategy;
+
+// Passport session setup.
+//   To support persistent login sessions, Passport needs to be able to
+//   serialize users into and deserialize users out of the session.  Typically,
+//   this will be as simple as storing the user ID when serializing, and finding
+//   the user by ID when deserializing.  However, since this example does not
+//   have a database of user records, the complete Goodreads profile is
+//   serialized and deserialized.
+passport.serializeUser(function(user, done) {
+  done(null, user);
+});
+
+passport.deserializeUser(function(obj, done) {
+  done(null, obj);
+});
+
+
+// Use the GoodreadsStrategy within Passport.
+//   Strategies in passport require a `verify` function, which accept
+//   credentials (in this case, a token, tokenSecret, and Goodreads profile), and
+//   invoke a callback with a user object.
+passport.use(new GoodreadsStrategy({
+    consumerKey: process.env.GOODREADS_KEY,
+    consumerSecret: process.env.GOODREADS_SECRET,
+    callbackURL: "http://localhost:3000/goodreads/callback"
+  },
+  function(token, tokenSecret, profile, done) {
+    // asynchronous verification, for effect...
+    process.nextTick(function () {
+      
+      // To keep the example simple, the user's Goodreads profile is returned to
+      // represent the logged-in user.  In a typical application, you would want
+      // to associate the Goodreads account with a user record in your database,
+      // and return that user instead.
+      return done(null, profile);
+    });
+  }
+));
 
 var app = module.exports = express();
 
-
-/**
- * Configuration
- */
-
-// all environments
+// Configuration
 app.set('port', process.env.PORT || 3000);
 app.set('views', __dirname + '/views');
 app.set('view engine', 'jade');
 app.use(morgan('dev'));
-app.use(bodyParser());
+app.use(bodyParser.json());
 app.use(methodOverride());
 app.use(express.static(path.join(__dirname, 'public')));
+
+app.use(session({ 
+    secret: 'keyboard cat',
+    resave: true,
+    saveUninitialized: true 
+    }));
+// Initialize Passport!  Also use passport.session() middleware, to support 
+// persistent login sessions (recommended). 
+app.use(passport.initialize()); 
+app.use(passport.session()); 
 
 var env = process.env.NODE_ENV || 'development';
 
@@ -40,25 +92,36 @@ if (env === 'production') {
   // TODO
 }
 
-/**
- * Routes
- */
-
+// Routes
 // serve index and view partials
 app.get('/', routes.index);
 app.get('/partials/:name', routes.partials);
 
-// JSON API
+// Api
 app.get('/api/name', api.name);
+app.get('/goodreads',
+    passport.authenticate('goodreads'),
+    function (req, res) {
+        // The request will be redirected to Goodreads for authentication, so this
+        // function will not be called.
+    });
 
-// redirect all others to the index (HTML5 history)
+// GET /auth/goodreads/callback
+//   Use passport.authenticate() as route middleware to authenticate the
+//   request.  If authentication fails, the user will be redirected back to the
+//   login page.  Otherwise, the primary route function function will be called,
+//   which, in this example, will redirect the user to the home page.
+app.get('/goodreads/callback',
+    passport.authenticate('goodreads', { failureRedirect: '/login' }),
+    function (req, res) {
+        console.log("Got callback from goodreads");
+        res.redirect('/');
+    });
+
+// Redirect all others to the index (HTML5 history)
 app.get('*', routes.index);
 
-
-/**
- * Start Server
- */
-
+// Start
 http.createServer(app).listen(app.get('port'), function () {
   console.log('Express server listening on port ' + app.get('port'));
 });
