--- bookman-old/.travis.yml
@@ -1,12 +1,0 @@
-language: node_js
-node_js:
-- '4.2'
-install:
-- npm install
-script:
-- npm run test-travis-1
-- npm run test-travis-2
-env:
-  global:
-  - secure: MpnDhnFnlH89HfB+4YNE0ly9THF53H7H6jX9F+uIV34kepKSmHufJtgJwKUdHB3tm+CIVyefjiE4Zd5aBODrOsG5sL1aO4ik4ZNuEVZOsZhQ4l3oQkFPu/zPhXBwOLGjIaXQNnekvKgmFvjFBIfSGFzdzvfc7ozhg3GxXTkrZeqBSFrgBKIO3OyzfRUci9lGkmXBfvVt9JqQPpx/bihjn59AP36NzyM32V2ZnjGkeuk60RYFrn0zy+0w15ajF1jBvxryUjVz+SyHh+MTnTn+Zqz8A7hqWa/Vtfxihy+3WVVw/Cx7ecubGsfHdUppiJLmnelBvJR+zbqXUkr88AKwWpXBm7gDeEosF6fHtOB8/HUQOk4DhGSvHGTecC6R3NLGgExNdGOv2n6bURUS9Kl2x6MxD7eETLWGX5Hk5dnQwDMaGu9lJXFTohFTifRskGkn5G4j+gV2+Ymdv2VvJCm6ldMVlG++ZIfxV2EonJlHRXFpGlcaJP21jjltSfgQUvNt/BWYpYGQmVzpIgQ3gVaE9cO27ht2b5HSecINVm3KgknZeISF29/vmFUojkMKkRPCN6n4N0PZgm/dNa2iaVuJZH+DPvwg6sH/sDpmxBzejkYadtXjigiSR2ILGrgguQD+CGC1dwejzwuWFHjIOqAmiaEvmIUkE+1PkBd/Q5PaI4A=
-  - secure: MJgdtCTBqrIS69P7Sa5uacVlqQWBTLnfpM3lDaKOF0ZE7IxgvNS2VyQLVRUmlatU0TV+vKYOknpug88jd9r3ed0f7ExK/FUD9Xa2Mz5vr1nId+Je6QocrQxvZyvzZoze3w7OSRfaNg7Ma8OEpSx68UcgdjF80LMvoC7HNbh6FUKzHfz90sxPnjHCyCHL+kNNC/6OFEQZ7kKP4DipB0yPlEuhHL1CzXC2kkNyzL6pRm0Kgd5/g61g2RUmeECyiQgRr6NqWZaZ5xxopMo+g8APp1t6HKLZBgFP5MZL94qOzzh52nr8Kk6zXQvt+l+h0t1gX4dGeiSlg4dYCiuutpO4IW5EnyZTfxS6s+yS8zIVV3Cng0iFkWrh0eX9Fq1nx2W5iTfePgBj9r6ttfXtkuX/JjbNSzjQNbKg08LxS1NdDUZPVczFmgR/Z+/xsFW69qLnh6KZVMvLloq+G98QfmXr9lA2bAjBLftWiWA6eZgQFx9Jxmikh5oe6gK3EH1T3fGARsi9xma9/1xA8rgbT/laRl1gyjbKK7AitGCTw29HFn7+L4AJRW9JeKP7KHIZzE8jyFKqexw2wRz8yC+Hb7EBvfivu2Q7YyVE2e0Ek/MptwJWux/dStwjBHE+tsLpxpmTrjzZ/iKLNB2epf0NSTqOnxGS2DLA7hPjagriSphFrJ8=
