--- bookman-old/README.md
@@ -1,5 +1,0 @@
-# bookman
-A book manager for reading buffs.
-
-[![Build Status](https://travis-ci.org/olivif/bookman.svg?branch=master)](https://travis-ci.org/olivif/bookman)
-[![Coverage Status](https://codecov.io/github/olivif/bookman/coverage.svg?precision=2)](https://codecov.io/github/olivif/bookman)
