--- bookman-old/package.json
@@ -1,48 +1,0 @@
-{
-    "name": "bookman",
-    "version": "0.0.0",
-    "description": "A book manager ",
-    "scripts": {
-        "start": "node ./bin/www",
-        "test": "mocha",
-        "test-travis-1": "istanbul cover ./node_modules/mocha/bin/_mocha -- -R spec",
-        "test-travis-2": "./node_modules/.bin/codecov"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/olivif/bookman.git"
-    },
-    "keywords": [
-        "bookman",
-        "book",
-        "books",
-        "manager",
-        "reading",
-        "tracker"
-    ],
-    "author": "Olivia Ifrim",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/olivif/bookman/issues"
-    },
-    "homepage": "https://github.com/olivif/bookman#readme",
-    "dependencies": {
-        "body-parser": "~1.13.2",
-        "cookie-parser": "~1.3.5",
-        "debug": "~2.2.0",
-        "express": "~4.13.1",
-        "express-session": "^1.13.0",
-        "jade": "~1.11.0",
-        "morgan": "~1.6.1",
-        "passport": "^0.3.2",
-        "passport-goodreads": "^1.0.0",
-        "serve-favicon": "~2.3.0",
-        "dotenv": "^1.2.0"
-    },
-    "devDependencies": {
-        "codecov": "^1.0.1",
-        "istanbul": "^0.4.2",
-        "mocha": "^2.3.4",
-        "should": "^8.1.1"
-    }
-}
