--- bookman-old/routes/index.js
@@ -1,55 +1,0 @@
-var express = require('express');
-var router = express.Router();
-var passport = require('passport');
-
-// Simple route middleware to ensure user is authenticated.
-//   Use this route middleware on any resource that needs to be protected.  If
-//   the request is authenticated (typically via a persistent login session),
-//   the request will proceed.  Otherwise, the user will be redirected to the
-//   login page.
-function ensureAuthenticated(req, res, next) {
-  if (req.isAuthenticated()) { return next(); }
-  res.redirect('/login');
-}
-
-router.get('/', function(req, res){
-  res.render('index', { user: req.user, title: "bookman" });
-});
-
-router.get('/account', ensureAuthenticated, function(req, res){
-  res.render('account', { user: req.user });
-});
-
-router.get('/login', function(req, res){
-  res.render('login', { user: req.user });
-});
-
-// GET /auth/goodreads
-//   Use passport.authenticate() as route middleware to authenticate the
-//   request.  The first step in Goodreads authentication will involve redirecting
-//   the user to goodreads.com.  After authorization, Goodreads will redirect the user
-//   back to this application at /auth/goodreads/callback
-router.get('/auth/goodreads',
-  passport.authenticate('goodreads'),
-  function(req, res){
-    // The request will be redirected to Goodreads for authentication, so this
-    // function will not be called.
-  });
-
-// GET /auth/goodreads/callback
-//   Use passport.authenticate() as route middleware to authenticate the
-//   request.  If authentication fails, the user will be redirected back to the
-//   login page.  Otherwise, the primary route function function will be called,
-//   which, in this example, will redirect the user to the home page.
-router.get('/auth/goodreads/callback', 
-  passport.authenticate('goodreads', { failureRedirect: '/login' }),
-  function(req, res) {
-    res.redirect('/');
-  });
-
-router.get('/logout', function(req, res){
-  req.logout();
-  res.redirect('/');
-});
-
-module.exports = router;
