--- bookman-old/views/account.jade
@@ -1,5 +1,0 @@
-extends layout
-
-block content
-  p #{user.id}
-  p #{user.displayName}
