--- bookman-old/views/index.jade
@@ -1,10 +1,0 @@
-extends layout
-
-block content
-    h1= title
-
-    if (!user) 
-        p Welcome! Please log in 
-        <a href="/auth/goodreads">Login with Goodreads</a>
-    else
-        p Welcome #{user.displayName}
