--- bookman-old/views/login.jade
@@ -1,4 +1,0 @@
-extends layout
-
-block content
-    <a href="/auth/goodreads">Login with Goodreads</a>
